Rename snowobs transform module to tableHelpers
