+++ src/services/snowobs/tableHelpers.ts
